Удалил лишний проект, залил фиксы, добавил возможность выбора квартала и обработчик на сервере, сохранение картинки(только тест пока что)

+++ OOP/OOPServer/Program.cs
@@ -35,10 +35,13 @@ namespace OOPServer
                     stream.Read(buffer_for_int, 0, 4);
                     MessageType messageType = (MessageType)BitConverter.ToInt32(buffer_for_int, 0);
 
+                    stream.Read(buffer_for_int, 0, 4);
+                    int quarter_index = BitConverter.ToInt32(buffer_for_int, 0);
+
                     if (messageType == MessageType.SendFile)
                     {
                         stream.Read(buffer_for_int, 0, 4);
-                        int branch_id = BitConverter.ToInt32(buffer_for_int, 0);
+                        int branch_index = BitConverter.ToInt32(buffer_for_int, 0);
 
                         List<byte> data = new List<byte>();
                         do
@@ -49,20 +52,23 @@ namespace OOPServer
                                 data.Add(buffer[i]);
                         } while (stream.DataAvailable);
 
-                        string file_name = "branch" + branch_id + ".dat";
+                        string file_name = "br" + branch_index + "q" + quarter_index + ".dat";
                         File.WriteAllBytes(file_name, data.ToArray());
                         Console.WriteLine("File " + file_name + " was received");
                     }
                     else if (messageType == MessageType.GetResult)
                     {
-                        string[] files = Directory.GetFiles(Directory.GetCurrentDirectory(), "branch?.dat");
-                        Main_Form main_form = new Main_Form(files);
-                        main_form.ShowDialog();
-                        byte[] file = File.ReadAllBytes(result_file_name);
-                        stream.Write(file, 0, file.Length);
-                        Thread.Sleep(200);
-                        Console.WriteLine("File result.dat was sent");
-                        File.Delete(result_file_name);
+                        string[] files = Directory.GetFiles(Directory.GetCurrentDirectory(), "br?q" + quarter_index + ".dat");
+                        if (files.Length > 0)
+                        {
+                            Main_Form main_form = new Main_Form(files);
+                            main_form.ShowDialog();
+                            byte[] file = File.ReadAllBytes(result_file_name);
+                            stream.Write(file, 0, file.Length);
+                            Thread.Sleep(200);
+                            Console.WriteLine("File result.dat was sent");
+                            File.Delete(result_file_name);
+                        }
                     }
                     else
                     {
